[release prep] Fix warnings in the SDKSampleApp
Fix some warnings in the SDKSampleApp

+ Fix some int / NSInteger conversions
+ Fix some printf format specifiers types

+++ SDKSampleApp/SDKSampleApp/STOauthLoginViewController.m
@@ -558,7 +558,7 @@
  * called.
  */
 -(void)configureServers:(UIPickerView *)pickerView {
-    int index = [pickerView selectedRowInComponent:0];
+    NSInteger index = [pickerView selectedRowInComponent:0];
     NSString *env = [NSString stringWithFormat:@"%@", ENVIRONMENTS[index]];
     self.selectedEnv = env;
     NSLog(@"%@", env);
